feat(game-logic) - partial rhythm system

+++ include/game/components/rhythm/timing.h
@@ -4,6 +4,6 @@ namespace Game::Rhythm
 {
     struct Timing
     {
-        int timing;
+        float timing;
     };
 }
